Fix: Smart Menu, Profile Completion, Team Carousel bugs
Smart Menu Block:
- Fixed submenu indicator icon not working (used same icon as dropdown)
- Walker now uses depth-based icon selection: dropdown for top-level,
  submenu indicator for nested items
- Removed dead mainMenuIcon attribute that only added CSS class

Profile Completion Block:
- Fixed "Hide on 100% Complete" toggle not working
- Cast round() result to int for strict comparison (=== 100)

Team Carousel Block:
- Fixed images not showing on frontend
- Corrected attribute name from imageId to imageID in edit.js

Documentation:
- Updated links from developer.wbcomdesigns.com/docs to docs.wbcomdesigns.com
- Fixed in 6 files: README, getting-started, video scripts, FAQ, emails, admin

+++ plugins/gutenberg/blocks/profile-completion/profile-completion.php
@@ -170,7 +170,8 @@ if ( ! function_exists( 'wbcom_essential_format_profile_progress' ) ) {
 		);
 
 		if ( $user_progress_arr['total_fields'] > 0 ) {
-			$profile_completion_percentage                   = round( ( $user_progress_arr['completed_fields'] * 100 ) / $user_progress_arr['total_fields'] );
+			// Cast to int because round() returns float, and render.php uses strict comparison (===).
+			$profile_completion_percentage                   = (int) round( ( $user_progress_arr['completed_fields'] * 100 ) / $user_progress_arr['total_fields'] );
 			$user_prgress_formatted['completion_percentage'] = $profile_completion_percentage;
 		}
